Unify the target of disk0 in lab8 and lab8_result
The commit 2d1c200 fixes the "No rule to make target 'disk0'" in
lab8_result, but incorrectly introduces the same change to the Makefile
in lab8 which already has a target for the directory.

This commit unifies both Makefiles to fix those "No rule to make target"
and "overriding recipe for target" messages once and for all.

Fixes: 2d1c200 ("Create the sfs root directory if missing")
Signed-off-by: Junjie Mao <eternal.n08@gmail.com>